modify user_id to open_id
