+++ app/Models/User.php
@@ -14,21 +14,6 @@ class User extends BaseModel
 		'password',
 		'member_expire_time'
 	];
-
-	/**
-	 * @return \Illuminate\Database\Eloquent\Relations\HasMany
-	 */
-	public function resume()
-	{
-		return $this->hasOne(Resume::class);
-	}
-
-	/**
-	 * @return \Illuminate\Database\Eloquent\Relations\BelongsToMany
-	 */
-	public function questions()
-	{
-		return $this->belongsToMany(Question::class, 'user_question');
-	}
+	
 	
 }
