+++ composer.json
@@ -7,6 +7,7 @@
     "require": {
         "php": ">=5.6.4",
         "dingo/api": "1.0.x@dev",
+        "doctrine/dbal": "^2.5",
         "iwanli/wxxcx": "1.*",
         "laravel/framework": "5.4.*",
         "laravel/tinker": "~1.0",
